% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{auth_me}
\alias{auth_me}
\title{Authentication}
\usage{
auth_me()
}
\value{
if you are you have set you API key or not
}
\description{
This function check if you are authorized or not
}
\examples{
auth_me()
}
\author{
Yassin Abdelhady
}
