% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{full_guardian_call}
\alias{full_guardian_call}
\title{The Guardian call}
\usage{
full_guardian_call(
  orderby = "newest",
  page_numbers = 1,
  page_size = 10,
  to_date = to,
  from_date = from,
  loop = FALSE
)
}
\value{
the full data from the guardian
}
\description{
This function retrieve the data from the guardian API plus the body and the tags of the article
}
\examples{
full_guardian_call()
}
\author{
Yassin Abdelhady
}
