% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{full_guardian_call_by_date}
\alias{full_guardian_call_by_date}
\title{The Guardian call}
\usage{
full_guardian_call_by_date(date = Sys.Date())
}
\arguments{
\item{date}{A character string indicating the date for articles to retrieve.DEFAULT today date}
}
\value{
the full data from the guardian
}
\description{
This function retrieve the data from the guardian API using a specific date plus the body and the tags of the article
}
\examples{

\dontrun{
  full_guardian_call_by_date()
}
}
\author{
Yassin Abdelhady
}
