% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{guardian_call_by_date}
\alias{guardian_call_by_date}
\title{The Guardian call}
\usage{
guardian_call_by_date(date = Sys.Date())
}
\arguments{
\item{date}{A character string indicating the date for articles to retrieve.DEFAULT today date}
}
\value{
the data from the guardian
}
\description{
This function retrieve the data from the guardian API using a specific date
}
\examples{

\dontrun{
  guardian_call_by_date(date = "2023-03-10")
}
}
\author{
Yassin Abdelhady
}
