% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{guardian_key}
\alias{guardian_key}
\title{Set the Guardian API key}
\usage{
guardian_key(key)
}
\arguments{
\item{key}{character string containing your API key.}
}
\value{
set your API key
}
\description{
This function sets your API key
}
\examples{
guardian_key("your_api_key_here")
}
\author{
Yassin Abdelhady
}
