% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setup.R
\name{show_my_key}
\alias{show_my_key}
\title{The Guardian API key}
\usage{
show_my_key()
}
\value{
show your API key
}
\description{
This function retrive your API key
}
\examples{
show_my_key()
}
\author{
Yassin Abdelhady
}
