% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{full_guardian_call}
\alias{full_guardian_call}
\title{The Guardian call}
\usage{
full_guardian_call(
  search_term = NULL,
  orderby = "newest",
  page_numbers = 1,
  page_size = 10,
  to_date = Sys.Date(),
  from_date = Sys.Date() - 7,
  loop = FALSE
)
}
\arguments{
\item{search_term}{A character string containing the query term.}

\item{orderby}{A character string indicating how the results should be ordered.}

\item{page_numbers}{An integer vector indicating the page numbers to retrieve.}

\item{page_size}{An integer indicating the number of results per page.}

\item{to_date}{A character string indicating the latest date for articles to retrieve.}

\item{from_date}{A character string indicating the earliest date for articles to retrieve.}

\item{loop}{A logical indicating whether to make multiple API calls to retrieve all available results.}
}
\value{
the full data from the guardian
}
\description{
This function retrieve the data from the guardian API plus the body and the tags of the article
}
\examples{

\dontrun{
  full_guardian_call(search_term = "football")
}
}
\author{
Yassin Abdelhady
}
